# openapi/api/v1/openapi.yaml
openapi: 3.0.1
info:
  title: "Ecosyste.ms: Archives"
  description: "An open API service for inspecting package archives and files from many open source software ecosystems. "
  contact:
    name: Ecosyste.ms
    email: support@ecosyste.ms
    url: https://ecosyste.ms 
  version: 1.0.0
  license:
    name: CC-BY-SA-4.0
    url: https://creativecommons.org/licenses/by-sa/4.0/
externalDocs:
  description: GitHub Repository
  url: https://github.com/ecosyste-ms/archives
servers:
- url: https://archives.ecosyste.ms/api/v1
paths:
  /archives/list:
    get:
      summary: list files in a package archive
      operationId: list
      parameters:
        - name: url
          in: query
          description: url to package archive
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /archives/contents:
    get:
      summary: get contents of a path from a package archive
      operationId: contents
      parameters:
        - name: url
          in: query
          description: url to package archive
          required: true
          schema:
            type: string
        - name: path
          in: query
          description: path to file or folder in package archive
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
  /archives/readme:
    get:
      summary: get readme from a package archive
      operationId: readme
      parameters:
        - name: url
          in: query
          description: url to package archive
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Readme'
  /archives/changelog:
    get:
      summary: get changelog from a package archive
      operationId: changelog
      parameters:
        - name: url
          in: query
          description: url to package archive
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Changelog'
components:
  schemas:
    Content:
      type: object
      properties:
        name:
          type: string
        directory:
          type: boolean
        contents:
          type: string
    Readme:
      type: object
      properties:
        name:
          type: string
        raw:
          type: string
        html:
          type: string
        plain:
          type: string
        extension:
          type: string
        language:
          type: string
        other_readme_files:
          type: array
          items:
            type: string
    Changelog:
      type: object
      properties:
        name:
          type: string
        raw:
          type: string
        html:
          type: string
        plain:
          type: string
        extension:
          type: string
        language:
          type: string